+++ apps/cookbook/tsconfig.json
@@ -27,7 +27,8 @@
       "@kirbydesign/designsystem/tabs": ["libs/designsystem/tabs/index.ts"],
       "@kirbydesign/designsystem/slide-button": ["libs/designsystem/slide-button/index.ts"],
       "@kirbydesign/designsystem/range": ["libs/designsystem/range/index.ts"],
-      "@kirbydesign/designsystem/popover": ["libs/designsystem/popover/index.ts"]
+      "@kirbydesign/designsystem/popover": ["libs/designsystem/popover/index.ts"],
+      "@kirbydesign/designsystem/loading-overlay": ["libs/designsystem/loading-overlay/index.ts"]
     },
     "target": "es2020"
   },
