+++ apps/flows/tsconfig.json
@@ -45,7 +45,8 @@
       "@kirbydesign/designsystem/tabs": ["libs/designsystem/tabs/index.ts"],
       "@kirbydesign/designsystem/slide-button": ["libs/designsystem/slide-button/index.ts"],
       "@kirbydesign/designsystem/range": ["libs/designsystem/range/index.ts"],
-      "@kirbydesign/designsystem/popover": ["libs/designsystem/popover/index.ts"]
+      "@kirbydesign/designsystem/popover": ["libs/designsystem/popover/index.ts"],
+      "@kirbydesign/designsystem/loading-overlay": ["libs/designsystem/loading-overlay/index.ts"]
     },
     "target": "es2020"
   },
