+++ libs/designsystem/loading-overlay/ng-package.json
@@ -1,0 +1,1 @@
+{}
