+++ libs/designsystem/loading-overlay/src/index.ts
@@ -1,0 +1,1 @@
+export * from './public_api';
