+++ libs/designsystem/loading-overlay/src/loading-overlay.component.html
