+++ libs/designsystem/loading-overlay/src/loading-overlay.service.ts
