+++ libs/designsystem/tsconfig.spec.json
@@ -24,7 +24,8 @@
       "@kirbydesign/designsystem/tabs": ["tabs/index.ts"],
       "@kirbydesign/designsystem/slide-button": ["slide-button/index.ts"],
       "@kirbydesign/designsystem/range": ["range/index.ts"],
-      "@kirbydesign/designsystem/popover": ["popover/index.ts"]
+      "@kirbydesign/designsystem/popover": ["popover/index.ts"],
+      "@kirbydesign/designsystem/loading-overlay": ["loading-overlay/index.ts"]
     }
   },
   "files": ["test.ts"],
